requirements.txt
# Core Streamlit dependencies
streamlit>=1.28.0
streamlit-option-menu
streamlit-lottie
streamlit-extras

# Document processing
python-docx
pypdf==4.2.0
pdfplumber
docx2txt

# Data processing
pandas
numpy
plotly
matplotlib
seaborn

# AI/ML dependencies
google-generativeai
openai
nltk
scikit-learn

# Utilities
pillow
python-dotenv
requests
sqlalchemy
openpyxl
reportlab

# PDF processing (lightweight)
pdfminer.six==20221105

# Note: Heavy dependencies removed for cloud deployment compatibility
# If you need OCR or browser automation, install locally:
# pdf2image, pytesseract, selenium, webdriver-manager, chromedriver-autoinstaller
